the fsolve version
